Switch to less insane keycodes for plover-hid
The original plover-hid tried to reuse the keycodes from the GeminiPR protocol
and had pretty weird key ordering, making the unnecessarily hard to
implement.

Switch to sequential numbering of all standard steno keys, and then add
the extra keys after those.

The ordering is the same as the ordering used by the TXBolt protocol.